Add requirements from OpenAI
Modified the parameters and function calls of gridworld according to the requirements from OpenAI for creating and in the future merging our environment with them. 
https://github.com/openai/gym/tree/master/gym/envs
