+++ gym-gridworld/gym_gridworld/__init__.py
@@ -1,0 +1,6 @@
+from gym.envs.registration import register
+
+register(
+    id='gridworld-v0',
+    entry_point='gym_gridworld.envs:GridWorldEnv',
+)
