+++ gym-gridworld/gym_gridworld/envs/__init__.py
@@ -1,0 +1,1 @@
+from gym_gridworld.envs.gridworld_env import GridWorldEnv
