+++ gym-gridworld/setup.py
@@ -1,0 +1,6 @@
+from setuptools import setup
+
+setup(name='gym_foo',
+      version='0.0.1',
+      install_requires=['gym']
+      )
